move osvscanner_internal to ci
