+++ internal/ci/fixtures/vulns/test-vuln-diff-a-b.json
