+++ internal/ci/fixtures/vulns/test-vuln-diff-b-a.json
