+++ internal/ci/fixtures/vulns/test-vuln-diff-b-c.json
