+++ internal/ci/fixtures/vulns/test-vuln-diff-c-b.json
