+++ internal/ci/vulnerability_result_diff.go
@@ -1,4 +1,4 @@
-package osvscanner_internal
+package ci
 
 import (
 	"github.com/google/osv-scanner/pkg/grouper"
