+++ internal/ci/vulnerability_result_diff_test.go
@@ -1,9 +1,9 @@
-package osvscanner_internal_test
+package ci_test
 
 import (
 	"testing"
 
-	"github.com/google/osv-scanner/internal/osvscanner_internal"
+	"github.com/google/osv-scanner/internal/ci"
 	"github.com/google/osv-scanner/internal/testutility"
 	"github.com/google/osv-scanner/pkg/models"
 )
@@ -56,7 +56,7 @@ func TestDiffVulnerabilityResults(t *testing.T) {
 		tt := tt
 		t.Run(tt.name, func(t *testing.T) {
 			t.Parallel()
-			got := osvscanner_internal.DiffVulnerabilityResults(tt.args.oldRes, tt.args.newRes)
+			got := ci.DiffVulnerabilityResults(tt.args.oldRes, tt.args.newRes)
 			testutility.AssertMatchFixtureJSON(t, tt.wantPath, got)
 		})
 	}
